feat: minecraft-essnetials workspace
chore: removed cli from workspace
