--- crate/cli/.gitignore
@@ -1,1 +1,0 @@
-/target
