--- crate/cli/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "cli"
-version = "0.1.0-alpha"
-edition = "2021"
-
-[dependencies]
-clap = { version = "4.5.4", features = ["derive"] }
-minecraft-essentials = { path = "../core" }
